+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 # Unreleased
 
+- added: Report legacy payment ID's using the new core `memo` transaction property.
 - removed: Only bundle the package in CJS-format, and drop the ESM output.
+- removed: Stop supporting undocumented `getTransaction` options.
 
 # 1.1.1 (2023-08-15)